refactor: Bundle parser-related logic in the parser generator controller
