+++ crates/codegen/runtime/generator/src/parser/keyword_scanner_definition.rs
@@ -4,8 +4,8 @@ use codegen_grammar::{
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 
-use crate::parser_definition::VersionQualityRangeVecExtensions;
-use crate::scanner_definition::ScannerDefinitionNodeExtensions;
+use crate::parser::parser_definition::VersionQualityRangeVecExtensions;
+use crate::parser::scanner_definition::ScannerDefinitionNodeExtensions;
 
 pub trait KeywordScannerDefinitionExtensions {
     fn to_scanner_code(&self) -> TokenStream;
