+++ crates/codegen/runtime/generator/src/parser/scanner_definition.rs
@@ -5,7 +5,7 @@ use inflector::Inflector;
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
 
-use crate::parser_definition::VersionQualityRangeVecExtensions;
+use crate::parser::parser_definition::VersionQualityRangeVecExtensions;
 
 pub trait ScannerDefinitionExtensions {
     fn to_scanner_code(&self) -> TokenStream;
